- screenshot automation works now

+++ tools/screenshots.sh
@@ -1,6 +1,16 @@
 #!/bin/bash
 
-declare -a emulators=("7intabled" "10intabled" "16by9phone")
+exitfn () {
+  trap SIGINT
+  stop_emulator
+  echo -e "${RED}Killing screenshot HTTP server with PID: $SERVER_PID${NC}"
+  kill $SERVER_PID
+  exit
+}
+
+trap "exitfn" INT
+
+declare -a emulators=("7intablet" "10intablet" "16by9phone")
 
 # ANSI color codes
 RED='\033[0;31m'
@@ -10,22 +20,33 @@ BLUE='\033[0;34m'
 NC='\033[0m' # No Color
 
 start_http_server() {
-  echo -e "${BLUE}Starting screenshot HTTP server for emulator: $CURRENT_EMULATOR${NC}"
-  EMULATOR_NAME="$CURRENT_EMULATOR" dart integration_test/screenshot_server.dart &
-  SERVER_PID=$!
+  if [[ $(curl -s http://localhost:3824/health) != "true" ]]; then
+    echo -e "${BLUE}Starting screenshot HTTP server for emulator: $CURRENT_EMULATOR${NC}"
+    EMULATOR_NAME="$CURRENT_EMULATOR" dart integration_test/screenshot_server.dart &
+    SERVER_PID=$!
   
-  server_started=false
-  while [[ "$server_started" != "true" ]]; do
-    sleep 1
-    server_started=$(curl -s http://localhost:7384/health)
-    echo -e "${YELLOW}Waiting for start...${NC}"
-  done
+    server_started=false
+    while [[ "$server_started" != "true" ]]; do
+      sleep 1
+      server_started=$(curl -s http://localhost:3824/health)
+      echo -e "${YELLOW}Waiting for start...${NC}"
+    done
+  fi
+
+  if [[ $SERVER_PID == '' ]]; then
+    echo -e "${YELLOW}Server seems to be running from previous run. This could be intended. If not, kill it manually.${NC}"
+    return
+  fi
 
   echo -e "${GREEN}HTTP server started with PID: $SERVER_PID${NC}"
 }
 
 stop_http_server() {
   echo -e "${RED}Stopping screenshot HTTP server...${NC}"
+  if [[ $SERVER_PID == '' ]]; then
+    echo -e "${YELLOW}Server seems to be running from previous run. This could be intended. If not, kill it manually.${NC}"
+    return
+  fi
   kill $SERVER_PID
 }
 
@@ -35,7 +56,7 @@ start_emulator() {
   nohup emulator -avd "$emulator_name" -no-audio -no-window &
 
   boot_completed=""
-  while [[ "$boot_completed" != "1" ]]; do
+  while [[ $boot_completed != "1" ]]; do
     boot_completed=$(adb shell getprop sys.boot_completed 2>/dev/null)
     echo -e "${YELLOW}Waiting for $emulator_name to boot...${NC}"
     sleep 5
@@ -44,13 +65,25 @@ start_emulator() {
 }
 
 run_tests() {
+  SERVER_IP=$(ifconfig | grep -Eo 'inet (addr:)?([0-9]*\.){3}[0-9]*' | grep -Eo '([0-9]*\.){3}[0-9]*' | grep -v '127.0.0.1' | tail -n 1)
+
+  echo -e "${BLUE}SERVER_IP: $SERVER_IP${NC}"
+
   echo -e "${BLUE}Starting Flutter testing...${NC}"
-  flutter test integration_test/screenshot_automation_test.dart -d "emulator"
+  flutter test integration_test/screenshot_automation_test.dart -d "emulator" --dart-define=SERVER_IP=$SERVER_IP
+
+  if [ $? -ne 0 ]; then
+    echo -e "${RED}Flutter tests failed.${NC}"
+    kill $SERVER_PID
+    stop_emulator
+    trap SIGINT
+    exit 1
+  fi
 }
 
 stop_emulator() {
-  echo -e "${RED}Stopping emulator${NC}"
-  adb emu kill
+  echo -e "${RED}Stopping all emulators${NC}"
+  adb emu kill -as
   sleep 5
 }
 
@@ -61,8 +94,11 @@ for emulator in "${emulators[@]}"; do
   start_emulator "$CURRENT_EMULATOR"
   run_tests "$CURRENT_EMULATOR"
   stop_emulator
+  sleep 5
   stop_http_server
   echo -e "${GREEN}Finished tests on $CURRENT_EMULATOR${NC}"
 done
 
 echo -e "${GREEN}All tests completed on all emulators.${NC}"
+
+trap SIGINT
